Add cargo apk install command
